+++ mx.h
@@ -179,7 +179,7 @@ struct MxOps
    * **Contract**
    * - @a m is not NULL
    */
-  int (*mbox_check_stats)(struct Mailbox *m, int flags);
+  int (*mbox_check_stats)(struct Mailbox *m, uint8_t flags);
 
   /**
    * mbox_sync - Save changes to the Mailbox
@@ -375,7 +375,7 @@ struct MxOps
 
 /* Wrappers for the Mailbox API, see MxOps */
 int             mx_mbox_check      (struct Mailbox *m);
-int             mx_mbox_check_stats(struct Mailbox *m, int flags);
+int             mx_mbox_check_stats(struct Mailbox *m, uint8_t flags);
 int             mx_mbox_close      (struct Context **ptr);
 struct Context *mx_mbox_open       (struct Mailbox *m, OpenMailboxFlags flags);
 int             mx_mbox_sync       (struct Mailbox *m);
